test_: add multiple status-backend instances
