+++ tests-functional/clients/status_backend.py
@@ -1,87 +1,38 @@
 import json
 import logging
 import time
-from datetime import datetime
-from json import JSONDecodeError
-
-import jsonschema
+import random
+import threading
 import requests
 
 from clients.signals import SignalClient
+from clients.rpc import RpcClient
+from datetime import datetime
 from conftest import option
 from constants import user_1
 
 
-class RpcClient:
-
-    def __init__(self, rpc_url, client=requests.Session()):
-        self.client = client
-        self.rpc_url = rpc_url
-
-    def _check_decode_and_key_errors_in_response(self, response, key):
-        try:
-            return response.json()[key]
-        except json.JSONDecodeError:
-            raise AssertionError(
-                f"Invalid JSON in response: {response.content}")
-        except KeyError:
-            raise AssertionError(
-                f"Key '{key}' not found in the JSON response: {response.content}")
-
-    def verify_is_valid_json_rpc_response(self, response, _id=None):
-        assert response.status_code == 200, f"Got response {response.content}, status code {response.status_code}"
-        assert response.content
-        self._check_decode_and_key_errors_in_response(response, "result")
-
-        if _id:
-            try:
-                if _id != response.json()["id"]:
-                    raise AssertionError(
-                        f"got id: {response.json()['id']} instead of expected id: {_id}"
-                    )
-            except KeyError:
-                raise AssertionError(f"no id in response {response.json()}")
-        return response
-
-    def verify_is_json_rpc_error(self, response):
-        assert response.status_code == 200
-        assert response.content
-        self._check_decode_and_key_errors_in_response(response, "error")
-
-    def rpc_request(self, method, params=[], request_id=13, url=None):
-        url = url if url else self.rpc_url
-        data = {"jsonrpc": "2.0", "method": method, "id": request_id}
-        if params:
-            data["params"] = params
-        logging.info(f"Sending POST request to url {url} with data: {json.dumps(data, sort_keys=True, indent=4)}")
-        response = self.client.post(url, json=data)
-        try:
-            logging.info(f"Got response: {json.dumps(response.json(), sort_keys=True, indent=4)}")
-        except JSONDecodeError:
-            logging.info(f"Got response: {response.content}")
-        return response
-
-    def rpc_valid_request(self, method, params=[], _id=None, url=None):
-        response = self.rpc_request(method, params, _id, url)
-        self.verify_is_valid_json_rpc_response(response, _id)
-        return response
-            
-    def verify_json_schema(self, response, method):
-        with open(f"{option.base_dir}/schemas/{method}", "r") as schema:
-            jsonschema.validate(instance=response,
-                                schema=json.load(schema))
-
 
 class StatusBackend(RpcClient, SignalClient):
 
-    def __init__(self, await_signals=list()):
+    def __init__(self, await_signals=[], url=None):
+        try:
+            url = url if url else random.choice(option.status_backend_urls)
+        except IndexError:
+            raise Exception("Not enough status-backend containers, please add more")
+        option.status_backend_urls.remove(url)
 
-        self.api_url = f"{option.rpc_url_status_backend}/statusgo"
-        self.ws_url = f"{option.ws_url_status_backend}"
-        self.rpc_url = f"{option.rpc_url_status_backend}/statusgo/CallRPC"
+        self.api_url = f"{url}/statusgo"
+        self.ws_url = f"{url}".replace("http", "ws")
+        self.rpc_url = f"{url}/statusgo/CallRPC"
+
 
         RpcClient.__init__(self, self.rpc_url)
         SignalClient.__init__(self, self.ws_url, await_signals)
+
+        websocket_thread = threading.Thread(target=self._connect)
+        websocket_thread.daemon = True
+        websocket_thread.start()
 
     def api_request(self, method, data, url=None):
         url = url if url else self.api_url
